food_stories/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.7
django-3-jet==1.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
djangorestframework==3.11.0
future==0.18.2
google-api-python-client==1.4.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
Markdown==3.2.2
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
